Modularize layout: extract repositionPlayers/updateGrimoire/positionRadialStack into ui/layout.js and wire calls; update SW cache
