+++ service-worker.js
@@ -13,6 +13,7 @@ const urlsToCache = [
   './ui/guides.js',
   './ui/sidebar.js',
   './ui/tour.js',
+  './ui/layout.js',
   './characters.json',
   './Trouble Brewing.json',
   './Bad Moon Rising.json',
@@ -34,7 +35,7 @@ const urlsToCache = [
 const networkFirstFiles = [
   'index.html',
   'styles.css',
-  'script.js', 'utils.js', 'pwa.js', 'ui/tooltip.js', 'ui/svg.js', 'ui/guides.js', 'ui/sidebar.js', 'ui/tour.js',
+  'script.js', 'utils.js', 'pwa.js', 'ui/tooltip.js', 'ui/svg.js', 'ui/guides.js', 'ui/sidebar.js', 'ui/tour.js', 'ui/layout.js',
   'service-worker.js',
   'manifest.json'
 ];
